Fix 3 warnings

+++ include/pstem/indexing/string_util.h
@@ -41,11 +41,11 @@ inline wchar_t tolower_western(const wchar_t c)
 // strtol
 inline double strtol(const char * str, char ** strend, int radix)
 {
-    return std::strtol(str, strend, radix);
+    return (double)std::strtol(str, strend, radix);
 }
 inline double strtol(const wchar_t * str, wchar_t ** strend, int radix)
 {
-    return std::wcstol(str, strend, radix);
+    return (double)std::wcstol(str, strend, radix);
 }
 
 // strtod
